pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "maplebot"
version = "1.0.0"
description = "GMSR 群机器人，基于 NoneBot2 + OneBot 11 协议"
requires-python = ">=3.12"
dependencies = [
    "nonebot2[websockets]>=2.4.0",
    "nonebot-plugin-apscheduler>=0.5.0",
    "nonebot-adapter-onebot>=2.4.0",
    "nonebot-adapter-console>=0.6.0",
    "httpx>=0.27.0",
    "dynaconf>=3.2.0",
    "PyYAML>=6.0",
    "matplotlib>=3.8.0",
    "numpy>=2.0.0",
    "jieba>=0.42.1",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pylint>=4.0.0",
    "build>=1.4.0",
]

[tool.setuptools.packages.find]
include = ["maplebot*"]
